Use parsePSLists and serializeTrie to generate tld lists
